Remove CRATER (Abandoned) (#79)
This software is no longer under development by its creators and is not in a stable state. The last release was over 1 year ago and the last commit was 6+ months ago. In the project's official discord, the maintainers have stated that they are only able to provide critical security updates (due to them working on a closed-source project) however, there are many bugs and missing functionality in the software that has caused people discovering it from this list to waste multiple hours to set it up, only to be disappointed. I can provide further info if needed.